refactor: split parser mod.rs into submodules
