+++ src/parser/mod.rs
@@ -1,16 +1,30 @@
-use crate::models::{
-    AssetLookupSource, Contract, Expression, Function, GroupIOSource, GroupSumSource, Parameter,
-    Requirement, Statement,
-};
+use crate::models::{Contract, Function, Parameter, Statement};
 use pest::iterators::{Pair, Pairs};
 use pest::Parser;
 use pest_derive::Parser;
-use std::str::FromStr;
 
 /// Pest parser generated from grammar.pest
 #[derive(Parser)]
 #[grammar = "parser/grammar.pest"]
 pub struct ArkadeParser;
+
+// Parser split into domain submodules. Each declares its pub(crate)
+// parse_* helpers; siblings reach them via `use super::*`.
+mod asset;
+mod checksig;
+mod comparison;
+mod crypto;
+mod expr;
+mod introspection;
+mod tapscript;
+
+pub(crate) use asset::*;
+pub(crate) use checksig::*;
+pub(crate) use comparison::*;
+pub(crate) use crypto::*;
+pub(crate) use expr::*;
+pub(crate) use introspection::*;
+pub(crate) use tapscript::*;
 
 /// Parse Arkade Script source code into a Contract AST.
 ///
@@ -133,201 +147,6 @@ fn parse_function(pair: Pair<Rule>) -> Result<Function, String> {
     }
 
     Ok(func)
-}
-
-/// True if a `function` pair carries a `tapscript_block` body.
-fn function_pair_is_tapscript(pair: &Pair<Rule>) -> bool {
-    pair.clone()
-        .into_inner()
-        .any(|p| p.as_rule() == Rule::tapscript_block)
-}
-
-/// Parse a `function <name>(<params>) tapscript { … }` declaration.
-fn parse_named_tapscript(pair: Pair<Rule>) -> Result<crate::models::NamedTapscript, String> {
-    let mut inner = pair.into_inner();
-    let name = inner
-        .next()
-        .ok_or("Missing tapscript name")?
-        .as_str()
-        .to_string();
-    let params = inner.next().ok_or("Missing tapscript parameter list")?;
-    let inputs = parse_parameters(params)?;
-    let block = inner.next().ok_or("Missing tapscript body")?;
-    let mut items = Vec::new();
-    for stmt in block.into_inner() {
-        if stmt.as_rule() == Rule::require_stmt {
-            let expr = stmt
-                .into_inner()
-                .next()
-                .ok_or("Empty require() in tapscript")?;
-            items.push(parse_tap_item(expr)?);
-        }
-    }
-    Ok(crate::models::NamedTapscript {
-        name,
-        inputs,
-        items,
-    })
-}
-
-/// Interpret one `require(...)` inner expression as a tapscript item.
-fn parse_tap_item(pair: Pair<Rule>) -> Result<crate::models::TapItem, String> {
-    use crate::models::{HashFn, TapItem};
-    match pair.as_rule() {
-        Rule::hash_comparison => {
-            let mut inner = pair.into_inner();
-            let func = inner.next().ok_or("Missing hash function")?; // hash_func
-            let mut f_inner = func.into_inner();
-            let fn_name = f_inner.next().ok_or("Missing hash fn name")?.as_str();
-            let hash_fn =
-                HashFn::parse(fn_name).ok_or_else(|| format!("unknown hash function {fn_name}"))?;
-            let preimage = f_inner
-                .next()
-                .ok_or("Missing hash preimage")?
-                .as_str()
-                .to_string();
-            let hash = inner
-                .next()
-                .ok_or("Missing hash value")?
-                .as_str()
-                .to_string();
-            Ok(TapItem::Hash {
-                hash_fn,
-                preimage,
-                hash,
-            })
-        }
-        Rule::time_comparison => {
-            // tx.time >= ident  → absolute (CLTV)
-            let mut inner = pair.into_inner();
-            let value = inner
-                .next()
-                .ok_or("Missing tx.time bound")?
-                .as_str()
-                .to_string();
-            Ok(TapItem::After { value })
-        }
-        Rule::check_sig => {
-            let mut inner = pair.into_inner();
-            let sig = inner
-                .next()
-                .ok_or("Missing signature")?
-                .as_str()
-                .to_string();
-            let key = parse_key_expr(inner.next().ok_or("Missing key")?)?;
-            Ok(TapItem::Sig {
-                keys: vec![key],
-                sigs: vec![sig],
-                threshold: Some(1),
-            })
-        }
-        Rule::check_multisig => {
-            // check_multisig wraps check_threshold_multisig.
-            let inner = pair
-                .into_inner()
-                .next()
-                .ok_or("Missing checkMultisig body")?;
-            parse_tap_multisig(inner)
-        }
-        Rule::function_call => {
-            // older(n) / after(n)
-            let mut inner = pair.into_inner();
-            let name = inner
-                .next()
-                .ok_or("Missing call name")?
-                .as_str()
-                .to_string();
-            let arg = inner
-                .next()
-                .ok_or_else(|| format!("{name}() requires one argument"))?
-                .as_str()
-                .to_string();
-            match name.as_str() {
-                "older" => {
-                    if inner.next().is_some() {
-                        return Err(format!("{name}() requires one argument"));
-                    }
-                    Ok(TapItem::Older { value: arg })
-                }
-                "after" => {
-                    if inner.next().is_some() {
-                        return Err(format!("{name}() requires one argument"));
-                    }
-                    Ok(TapItem::After { value: arg })
-                }
-                other => Err(format!("unsupported tapscript call `{other}(...)`")),
-            }
-        }
-        other => Err(format!(
-            "unsupported expression in tapscript require(): {other:?}"
-        )),
-    }
-}
-
-/// Parse `check_threshold_multisig` inner pairs into a Sig item.
-fn parse_tap_multisig(pair: Pair<Rule>) -> Result<crate::models::TapItem, String> {
-    use crate::models::TapItem;
-    let mut keys = Vec::new();
-    let mut sigs = Vec::new();
-    let mut threshold = None;
-    for child in pair.into_inner() {
-        match child.as_rule() {
-            Rule::key_array => {
-                for k in child.into_inner() {
-                    keys.push(parse_key_expr(k)?);
-                }
-            }
-            Rule::array => {
-                for s in child.into_inner() {
-                    sigs.push(s.as_str().to_string());
-                }
-            }
-            Rule::number_literal => {
-                threshold = Some(
-                    child
-                        .as_str()
-                        .parse::<u16>()
-                        .map_err(|e| format!("invalid threshold: {e}"))?,
-                );
-            }
-            _ => {}
-        }
-    }
-    Ok(TapItem::Sig {
-        keys,
-        sigs,
-        threshold,
-    })
-}
-
-/// Parse a `key_expr` (bare identifier or `tweak(emulator, func)`).
-fn parse_key_expr(pair: Pair<Rule>) -> Result<crate::models::KeyExpr, String> {
-    use crate::models::KeyExpr;
-    match pair.as_rule() {
-        Rule::identifier => Ok(KeyExpr::Ident(pair.as_str().to_string())),
-        Rule::key_expr => {
-            let inner = pair.into_inner().next().ok_or("Empty key expression")?;
-            parse_key_expr(inner)
-        }
-        Rule::tweak_key => {
-            let mut inner = pair.into_inner();
-            let base = inner
-                .next()
-                .ok_or("Missing tweak base")?
-                .as_str()
-                .to_string();
-            if base != "emulator" {
-                return Err(format!("tweak() base must be `emulator`, got `{base}`"));
-            }
-            let func = inner
-                .next()
-                .ok_or("Missing tweak func name")?
-                .as_str()
-                .to_string();
-            Ok(KeyExpr::Tweak { func })
-        }
-        other => Err(format!("unexpected key expression: {other:?}")),
-    }
 }
 
 /// Parse a statement in a function body (require, let binding, function call, variable declaration)
@@ -484,2170 +303,8 @@ fn parse_block(pair: Pair<Rule>) -> Result<Vec<Statement>, String> {
     Ok(statements)
 }
 
-// Parse general expression (with operator precedence)
-fn parse_general_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    match pair.as_rule() {
-        Rule::general_expression | Rule::comparison_expr => {
-            // Unwrap and parse the inner expression
-            let mut inner = pair.into_inner();
-            if let Some(first) = inner.next() {
-                let left = parse_additive_expr(first)?;
-
-                // Check for comparison operator
-                if let Some(op_pair) = inner.next() {
-                    let op = op_pair.as_str().to_string();
-                    let right_pair = inner.next().ok_or("Missing right side of comparison")?;
-                    let right = parse_additive_expr(right_pair)?;
-                    Ok(Expression::BinaryOp {
-                        left: Box::new(left),
-                        op,
-                        right: Box::new(right),
-                    })
-                } else {
-                    Ok(left)
-                }
-            } else {
-                Err("Empty expression".to_string())
-            }
-        }
-        Rule::additive_expr => parse_additive_expr(pair),
-        Rule::multiplicative_expr => parse_multiplicative_expr(pair),
-        Rule::unary_expr | Rule::primary_expr => parse_primary_expr(pair),
-        Rule::identifier => Ok(Expression::Variable(pair.as_str().to_string())),
-        Rule::number_literal => Ok(Expression::Literal(pair.as_str().to_string())),
-        Rule::tx_property_access => parse_tx_property_to_expr(pair),
-        Rule::this_property_access => Ok(Expression::Property(pair.as_str().to_string())),
-        _ => {
-            // Try to parse as a primary expression
-            parse_primary_expr(pair)
-        }
-    }
-}
-
-// Parse additive expression (+ and -)
-fn parse_additive_expr(pair: Pair<Rule>) -> Result<Expression, String> {
-    match pair.as_rule() {
-        Rule::additive_expr => {
-            let mut inner = pair.into_inner();
-            let first = inner
-                .next()
-                .ok_or("Missing first operand in additive expression")?;
-            let mut result = parse_multiplicative_expr(first)?;
-
-            // Process remaining operands
-            while let Some(op_pair) = inner.next() {
-                let op = op_pair.as_str().to_string();
-                let right_pair = inner
-                    .next()
-                    .ok_or("Missing right operand in additive expression")?;
-                let right = parse_multiplicative_expr(right_pair)?;
-                result = Expression::BinaryOp {
-                    left: Box::new(result),
-                    op,
-                    right: Box::new(right),
-                };
-            }
-
-            Ok(result)
-        }
-        _ => parse_multiplicative_expr(pair),
-    }
-}
-
-// Parse multiplicative expression (* and /)
-fn parse_multiplicative_expr(pair: Pair<Rule>) -> Result<Expression, String> {
-    match pair.as_rule() {
-        Rule::multiplicative_expr => {
-            let mut inner = pair.into_inner();
-            let first = inner
-                .next()
-                .ok_or("Missing first operand in multiplicative expression")?;
-            let mut result = parse_primary_expr(first)?;
-
-            // Process remaining operands
-            while let Some(op_pair) = inner.next() {
-                let op = op_pair.as_str().to_string();
-                let right_pair = inner
-                    .next()
-                    .ok_or("Missing right operand in multiplicative expression")?;
-                let right = parse_primary_expr(right_pair)?;
-                result = Expression::BinaryOp {
-                    left: Box::new(result),
-                    op,
-                    right: Box::new(right),
-                };
-            }
-
-            Ok(result)
-        }
-        _ => parse_primary_expr(pair),
-    }
-}
-
-// Parse primary expression (atoms)
-fn reject_reserved_function_call(pair: &Pair<Rule>) -> Result<(), String> {
-    if pair.as_rule() != Rule::function_call {
-        return Ok(());
-    }
-
-    let name = pair
-        .clone()
-        .into_inner()
-        .next()
-        .ok_or("Missing call name")?
-        .as_str()
-        .to_string();
-
-    if let Some(signature) = reserved_function_signature(&name) {
-        return Err(format!(
-            "malformed reserved function call `{name}(...)`; expected {signature}"
-        ));
-    }
-
-    Ok(())
-}
-
-fn reserved_function_signature(name: &str) -> Option<&'static str> {
-    match name {
-        "checkSig" => Some("checkSig(signature, pubkey)"),
-        "checkSigFromStack" => Some("checkSigFromStack(signature, pubkey, message)"),
-        "checkSigFromStackVerify" => Some("checkSigFromStackVerify(signature, pubkey, message)"),
-        "checkMultisig" => Some("checkMultisig([pubkeys], [sigs]?, threshold?)"),
-        "sha256" => Some("sha256(data)"),
-        "hash160" => Some("hash160(data)"),
-        "hash256" => Some("hash256(data)"),
-        "ripemd160" => Some("ripemd160(data)"),
-        "sha256Initialize" => Some("sha256Initialize(data)"),
-        "sha256Update" => Some("sha256Update(ctx, chunk)"),
-        "sha256Finalize" => Some("sha256Finalize(ctx, lastChunk)"),
-        "neg64" => Some("neg64(value)"),
-        "le64ToScriptNum" => Some("le64ToScriptNum(value)"),
-        "le32ToLe64" => Some("le32ToLe64(value)"),
-        "ecMulScalarVerify" => Some("ecMulScalarVerify(k, P, Q)"),
-        "tweakVerify" => Some("tweakVerify(P, k, Q)"),
-        "older" => Some("older(value)"),
-        "after" => Some("after(value)"),
-        _ => None,
-    }
-}
-
-fn parse_primary_expr(pair: Pair<Rule>) -> Result<Expression, String> {
-    match pair.as_rule() {
-        Rule::primary_expr | Rule::unary_expr => {
-            let inner = pair.into_inner().next().ok_or("Empty primary expression")?;
-            parse_primary_expr(inner)
-        }
-        Rule::general_expression | Rule::comparison_expr => {
-            // Parenthesized expression
-            parse_general_expression(pair)
-        }
-        Rule::identifier => Ok(Expression::Variable(pair.as_str().to_string())),
-        Rule::number_literal => Ok(Expression::Literal(pair.as_str().to_string())),
-        Rule::tx_property_access => parse_tx_property_to_expr(pair),
-        Rule::this_property_access => Ok(Expression::Property(pair.as_str().to_string())),
-        Rule::check_sig => {
-            let mut inner = pair.into_inner();
-            let signature = inner
-                .next()
-                .ok_or("Missing signature")?
-                .as_str()
-                .to_string();
-            let pubkey = inner.next().ok_or("Missing pubkey")?.as_str().to_string();
-            Ok(Expression::CheckSigExpr { signature, pubkey })
-        }
-        Rule::check_sig_from_stack => {
-            let mut inner = pair.into_inner();
-            let signature = inner
-                .next()
-                .ok_or("Missing signature")?
-                .as_str()
-                .to_string();
-            let pubkey = inner.next().ok_or("Missing pubkey")?.as_str().to_string();
-            let message = inner.next().ok_or("Missing message")?.as_str().to_string();
-            Ok(Expression::CheckSigFromStackExpr {
-                signature,
-                pubkey,
-                message,
-            })
-        }
-        Rule::sha256_func => {
-            // sha256(data) → one-shot OP_SHA256 over the inner expression.
-            let inner = pair.into_inner().next().ok_or("Missing sha256 argument")?;
-            let data = parse_additive_expr(inner)?;
-            Ok(Expression::Sha256 {
-                data: Box::new(data),
-            })
-        }
-        // Streaming SHA256
-        Rule::sha256_initialize => parse_sha256_initialize(pair),
-        Rule::sha256_update => parse_sha256_update(pair),
-        Rule::sha256_finalize => parse_sha256_finalize(pair),
-        // Conversion & Arithmetic
-        Rule::neg64_func => parse_neg64(pair),
-        Rule::le64_to_script_num => parse_le64_to_script_num(pair),
-        Rule::le32_to_le64 => parse_le32_to_le64(pair),
-        // Crypto Opcodes
-        Rule::ec_mul_scalar_verify => parse_ec_mul_scalar_verify(pair),
-        Rule::tweak_verify => parse_tweak_verify(pair),
-        Rule::check_sig_from_stack_verify => parse_check_sig_from_stack_verify_expr(pair),
-        // Byte-string manipulation
-        Rule::substr_func => parse_substr(pair),
-        Rule::cat_func => parse_cat(pair),
-        Rule::bin2num_func => parse_bin2num(pair),
-        Rule::num2bin_func => parse_num2bin(pair),
-        Rule::size_func => parse_size(pair),
-        // Packet introspection
-        Rule::packet_inspect => parse_packet_inspect(pair),
-        Rule::input_packet_inspect => parse_input_packet_inspect(pair),
-        Rule::asset_lookup => parse_asset_lookup_to_expression(pair),
-        Rule::asset_has => parse_asset_has_to_expression(pair),
-        Rule::asset_count => parse_asset_count_to_expression(pair),
-        Rule::asset_at => parse_asset_at_to_expression(pair),
-        Rule::group_control_is => parse_group_control_is_to_expression(pair),
-        Rule::input_introspection => parse_input_introspection_to_expression(pair),
-        Rule::output_introspection => parse_output_introspection_to_expression(pair),
-        Rule::tx_introspection => parse_tx_introspection_to_expression(pair),
-        Rule::constructor => parse_constructor_to_expression(pair),
-        Rule::function_call => {
-            reject_reserved_function_call(&pair)?;
-            Ok(Expression::Property(pair.as_str().to_string()))
-        }
-        Rule::additive_expr => parse_additive_expr(pair),
-        Rule::multiplicative_expr => parse_multiplicative_expr(pair),
-        _ => {
-            // Default to treating as a property string
-            Ok(Expression::Property(pair.as_str().to_string()))
-        }
-    }
-}
-
-/// Parse a complex expression into a Requirement AST node
-fn parse_complex_expression(pair: Pair<Rule>) -> Result<Requirement, String> {
-    match pair.as_rule() {
-        Rule::check_sig => parse_check_sig(pair),
-        Rule::check_sig_from_stack => parse_check_sig_from_stack(pair),
-        Rule::check_multisig => parse_check_multisig(pair),
-        Rule::time_comparison => parse_time_comparison(pair),
-        Rule::identifier_comparison => parse_identifier_comparison(pair),
-        Rule::property_comparison => parse_property_comparison(pair),
-        Rule::hash_comparison => parse_hash_comparison(pair),
-        Rule::byte_expr_comparison => parse_byte_expr_comparison(pair),
-        Rule::binary_operation => parse_binary_operation(pair),
-        Rule::asset_lookup_comparison => parse_asset_lookup_comparison(pair),
-        Rule::asset_count_comparison => parse_asset_count_comparison(pair),
-        Rule::asset_has_comparison => parse_asset_has_comparison(pair),
-        Rule::group_control_is_comparison => parse_group_control_is_comparison(pair),
-        Rule::asset_at_comparison => parse_asset_at_comparison(pair),
-        Rule::input_introspection_comparison => parse_input_introspection_comparison(pair),
-        Rule::output_introspection_comparison => parse_output_introspection_comparison(pair),
-        Rule::tx_introspection_comparison => parse_tx_introspection_comparison(pair),
-        Rule::input_introspection => parse_standalone_input_introspection(pair),
-        Rule::output_introspection => parse_standalone_output_introspection(pair),
-        Rule::tx_introspection => parse_standalone_tx_introspection(pair),
-        Rule::asset_lookup => parse_standalone_asset_lookup(pair),
-        Rule::asset_has => parse_standalone_asset_has(pair),
-        Rule::asset_count => parse_standalone_asset_count(pair),
-        Rule::asset_at => parse_standalone_asset_at(pair),
-        Rule::asset_group_access => parse_asset_group_access(pair),
-        Rule::group_control_is => parse_standalone_group_control_is(pair),
-        Rule::group_property_comparison => parse_group_property_comparison(pair),
-        // Streaming SHA256
-        Rule::sha256_initialize => {
-            let expr = parse_sha256_initialize(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::sha256_update => {
-            let expr = parse_sha256_update(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::sha256_finalize => {
-            let expr = parse_sha256_finalize(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        // Conversion & Arithmetic
-        Rule::neg64_func => {
-            let expr = parse_neg64(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::le64_to_script_num => {
-            let expr = parse_le64_to_script_num(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::le32_to_le64 => {
-            let expr = parse_le32_to_le64(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        // Crypto Opcodes
-        Rule::ec_mul_scalar_verify => {
-            let expr = parse_ec_mul_scalar_verify(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::tweak_verify => {
-            let expr = parse_tweak_verify(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::check_sig_from_stack_verify => parse_check_sig_from_stack_verify(pair),
-        // Byte-string manipulation — wrap as truthy assertions in require contexts.
-        Rule::substr_func => {
-            let expr = parse_substr(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::cat_func => {
-            let expr = parse_cat(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::bin2num_func => {
-            let expr = parse_bin2num(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::num2bin_func => {
-            let expr = parse_num2bin(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::size_func => {
-            let expr = parse_size(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::packet_inspect => {
-            let expr = parse_packet_inspect(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::input_packet_inspect => {
-            let expr = parse_input_packet_inspect(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::constructor => {
-            let expr = parse_constructor_to_expression(pair)?;
-            Ok(Requirement::Comparison {
-                left: expr,
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::tx_property_access | Rule::this_property_access => {
-            parse_property_access_as_requirement(pair)
-        }
-        Rule::function_call => {
-            reject_reserved_function_call(&pair)?;
-            let function_call = pair.as_str().to_string();
-            Ok(Requirement::Comparison {
-                left: Expression::Property(function_call),
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::identifier => {
-            let identifier = pair.as_str().to_string();
-            Ok(Requirement::Comparison {
-                left: Expression::Variable(identifier),
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        Rule::array_literal => {
-            let array_literal = pair.as_str().to_string();
-            Ok(Requirement::Comparison {
-                left: Expression::Property(array_literal),
-                op: "==".to_string(),
-                right: Expression::Literal("true".to_string()),
-            })
-        }
-        _ => Err(format!(
-            "Unexpected rule in complex expression: {:?}",
-            pair.as_rule()
-        )),
-    }
-}
-
-/// Parse checkSig(sig, pubkey) → CheckSig requirement
-fn parse_check_sig(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let signature = inner
-        .next()
-        .ok_or("Missing signature")?
-        .as_str()
-        .to_string();
-    let pubkey = inner
-        .next()
-        .ok_or("Missing public key")?
-        .as_str()
-        .to_string();
-    Ok(Requirement::CheckSig { signature, pubkey })
-}
-
-/// Parse checkSigFromStack(sig, pubkey, message) → CheckSigFromStack requirement
-fn parse_check_sig_from_stack(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let signature = inner
-        .next()
-        .ok_or("Missing signature")?
-        .as_str()
-        .to_string();
-    let pubkey = inner
-        .next()
-        .ok_or("Missing public key")?
-        .as_str()
-        .to_string();
-    let message = inner.next().ok_or("Missing message")?.as_str().to_string();
-    Ok(Requirement::CheckSigFromStack {
-        signature,
-        pubkey,
-        message,
-    })
-}
-
-/// Parse checkMultisig([pubkeys], [sigs]?, threshold?) → CheckMultisig requirement
-fn parse_check_multisig(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair
-        .into_inner()
-        .next()
-        .ok_or("Missing checkMultisig definition")?
-        .into_inner();
-    let pubkeys_array = inner.next().ok_or("Missing public keys")?;
-
-    let pubkeys: Vec<String> = pubkeys_array
-        .into_inner()
-        .map(|p| p.as_str().to_string())
-        .collect();
-
-    // Next may be an optional sigs array (Rule::array) or a threshold number.
-    // Skip the sigs array if present; use number_literal as threshold.
-    let mut threshold_pair = inner.next();
-    if let Some(ref tp) = threshold_pair {
-        if tp.as_rule() == Rule::array {
-            threshold_pair = inner.next();
-        }
-    }
-
-    match threshold_pair {
-        Some(next_pair) => {
-            // m-of-n threshold multisig
-            let threshold = match u16::from_str(next_pair.as_str()) {
-                Ok(threshold) => threshold,
-                Err(e) => return Err(format!("{}", e)),
-            };
-            Ok(Requirement::CheckMultisig { pubkeys, threshold })
-        }
-        None => {
-            // n-of-n multisig when threshold is omitted
-            let threshold = pubkeys.len() as u16;
-            Ok(Requirement::CheckMultisig { pubkeys, threshold })
-        }
-    }
-}
-
-/// Parse tx.time >= variable → After requirement
-fn parse_time_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let timelock_var = inner.next().ok_or("Missing timelock")?.as_str().to_string();
-    Ok(Requirement::After {
-        blocks: 0,
-        timelock_var: Some(timelock_var),
-    })
-}
-
-/// Parse identifier op identifier → After or Comparison requirement
-fn parse_identifier_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let left = inner
-        .next()
-        .ok_or("Missing left side expression")?
-        .as_str()
-        .to_string();
-    let op = inner
-        .next()
-        .ok_or("Missing comparison opcode")?
-        .as_str()
-        .to_string();
-    let right = inner
-        .next()
-        .ok_or("Missing right side expression")?
-        .as_str()
-        .to_string();
-
-    // Special case for time comparisons
-    if left == "tx.time" && op == ">=" {
-        return Ok(Requirement::After {
-            blocks: 0,
-            timelock_var: Some(right),
-        });
-    }
-
-    Ok(Requirement::Comparison {
-        left: Expression::Variable(left),
-        op,
-        right: Expression::Variable(right),
-    })
-}
-
-/// Parse property comparison: tx_property op expression
-fn parse_property_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let left_expr = inner.next().ok_or("Missing left side expression")?;
-    let op = inner
-        .next()
-        .ok_or("Missing comparison opcode")?
-        .as_str()
-        .to_string();
-    let right_expr = inner.next().ok_or("Missing right side expression")?;
-
-    reject_malformed_asset_call(left_expr.as_str())?;
-    reject_malformed_asset_call(right_expr.as_str())?;
-
-    let left = match left_expr.as_rule() {
-        Rule::tx_property_access | Rule::this_property_access => {
-            parse_tx_property_to_expression(left_expr)
-        }
-        _ => return Err("Unexpected left expression in property comparison".to_string()),
-    };
-
-    let right = match right_expr.as_rule() {
-        Rule::identifier => Expression::Variable(right_expr.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_expr.as_str().to_string()),
-        Rule::tx_property_access | Rule::this_property_access => {
-            parse_tx_property_to_expression(right_expr)
-        }
-        Rule::constructor => parse_constructor_to_expression(right_expr)?,
-        Rule::asset_lookup => parse_asset_lookup_to_expression(right_expr)?,
-        _ => return Err("Unexpected right expression in property comparison".to_string()),
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse sha256(preimage) == hash → HashEqual requirement
-fn parse_hash_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let hash_func = inner.next().ok_or("Missing hash function")?;
-    let mut hash_func_inner = hash_func.into_inner();
-    let fn_name = hash_func_inner
-        .next()
-        .ok_or("Missing hash function name")?
-        .as_str();
-    let hash_fn = crate::models::HashFn::parse(fn_name)
-        .ok_or_else(|| format!("unknown hash function {fn_name}"))?;
-    let preimage_pair = hash_func_inner.next().ok_or("Missing preimage")?;
-    let rhs_pair = inner.next().ok_or("Missing the hash")?;
-
-    // The grammar wraps the hash argument in `additive_expr`, so identifiers
-    // and literals surface as `Variable` / `Literal`, while byte-producing
-    // primitives (substr/cat/…) and arithmetic surface as their own variants.
-    let preimage_expr = parse_additive_expr(preimage_pair)?;
-    let rhs_is_identifier = matches!(rhs_pair.as_rule(), Rule::identifier);
-
-    // Fast path: a bare identifier/literal preimage AND identifier RHS keep the
-    // structured HashEqual emission (`<preimage> OP_<HASH> <hash> OP_EQUAL`).
-    if rhs_is_identifier {
-        if let Expression::Variable(name) | Expression::Literal(name) = &preimage_expr {
-            return Ok(Requirement::HashEqual {
-                hash_fn,
-                preimage: name.clone(),
-                hash: rhs_pair.as_str().to_string(),
-            });
-        }
-    }
-
-    // Complex preimage and/or complex RHS: emit via Comparison so byte-producing
-    // primitives expand inline. Only sha256 supports byte-expression operands.
-    if !matches!(hash_fn, crate::models::HashFn::Sha256) {
-        return Err(format!(
-            "{fn_name} with byte-expression operands is not supported; \
-             only sha256 allows substr/cat operands"
-        ));
-    }
-    let rhs_expr = match rhs_pair.as_rule() {
-        Rule::substr_func => parse_substr(rhs_pair)?,
-        Rule::cat_func => parse_cat(rhs_pair)?,
-        Rule::num2bin_func => parse_num2bin(rhs_pair)?,
-        Rule::identifier => Expression::Variable(rhs_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(rhs_pair.as_str().to_string()),
-        _ => Expression::Property(rhs_pair.as_str().to_string()),
-    };
-
-    Ok(Requirement::Comparison {
-        left: Expression::Sha256 {
-            data: Box::new(preimage_expr),
-        },
-        op: "==".to_string(),
-        right: rhs_expr,
-    })
-}
-
-/// Parse a `byte_expr_term` rule into an Expression
-/// (sha256/substr/cat/bin2num/num2bin/size).
-fn parse_byte_expr_term(pair: Pair<Rule>) -> Result<Expression, String> {
-    // byte_expr_term is a single-alternative wrapper — descend into the inner rule.
-    let inner = pair.into_inner().next().ok_or("Empty byte_expr_term")?;
-    match inner.as_rule() {
-        Rule::sha256_func => {
-            // sha256_func's child is always an additive_expr (see grammar), so
-            // route it through parse_additive_expr to recover the structured
-            // inner expression (substr/packet/cat/…) and wrap with
-            // Expression::Sha256 so the compiler emits inline OP_SHA256.
-            let arg_pair = inner.into_inner().next().ok_or("Missing sha256 argument")?;
-            let data = parse_additive_expr(arg_pair)?;
-            Ok(Expression::Sha256 {
-                data: Box::new(data),
-            })
-        }
-        Rule::substr_func => parse_substr(inner),
-        Rule::cat_func => parse_cat(inner),
-        Rule::bin2num_func => parse_bin2num(inner),
-        Rule::num2bin_func => parse_num2bin(inner),
-        Rule::size_func => parse_size(inner),
-        r => Err(format!("Unsupported byte_expr_term rule: {:?}", r)),
-    }
-}
-
-/// Parse a `byte_expr_atom` rule (one operand of byte_expr_arith).
-fn parse_byte_expr_atom(pair: Pair<Rule>) -> Result<Expression, String> {
-    let inner = pair.into_inner().next().ok_or("Empty byte_expr_atom")?;
-    match inner.as_rule() {
-        Rule::bin2num_func => parse_bin2num(inner),
-        Rule::size_func => parse_size(inner),
-        Rule::identifier => Ok(Expression::Variable(inner.as_str().to_string())),
-        Rule::number_literal => Ok(Expression::Literal(inner.as_str().to_string())),
-        r => Err(format!("Unsupported byte_expr_atom rule: {:?}", r)),
-    }
-}
-
-/// Parse a `byte_expr_arith` rule into a BinaryOp.
-fn parse_byte_expr_arith(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let left = parse_byte_expr_atom(inner.next().ok_or("Missing left of byte_expr_arith")?)?;
-    let op = inner
-        .next()
-        .ok_or("Missing op in byte_expr_arith")?
-        .as_str()
-        .to_string();
-    let right = parse_byte_expr_atom(inner.next().ok_or("Missing right of byte_expr_arith")?)?;
-    Ok(Expression::BinaryOp {
-        left: Box::new(left),
-        op,
-        right: Box::new(right),
-    })
-}
-
-/// Parse a `byte_expr_rhs` rule into an Expression.
-fn parse_byte_expr_rhs(pair: Pair<Rule>) -> Result<Expression, String> {
-    let inner = pair.into_inner().next().ok_or("Empty byte_expr_rhs")?;
-    match inner.as_rule() {
-        Rule::byte_expr_arith => parse_byte_expr_arith(inner),
-        Rule::byte_expr_term => parse_byte_expr_term(inner),
-        Rule::identifier => Ok(Expression::Variable(inner.as_str().to_string())),
-        Rule::number_literal => Ok(Expression::Literal(inner.as_str().to_string())),
-        r => Err(format!("Unsupported byte_expr_rhs rule: {:?}", r)),
-    }
-}
-
-/// Parse `byte_expr_comparison` → Requirement::Comparison.
-fn parse_byte_expr_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let left = parse_byte_expr_term(inner.next().ok_or("Missing left of byte_expr_comparison")?)?;
-    let op = inner
-        .next()
-        .ok_or("Missing op in byte_expr_comparison")?
-        .as_str()
-        .to_string();
-    let right = parse_byte_expr_rhs(
-        inner
-            .next()
-            .ok_or("Missing right of byte_expr_comparison")?,
-    )?;
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse binary operation: expr op expr → Comparison requirement
-fn parse_binary_operation(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let left_expr = inner.next().ok_or("Missing left side expression")?;
-    let op = inner
-        .next()
-        .ok_or("Missing binary opcode")?
-        .as_str()
-        .to_string();
-    let right_expr = inner.next().ok_or("Missing right side expression")?;
-
-    let left = match left_expr.as_rule() {
-        Rule::identifier => Expression::Variable(left_expr.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(left_expr.as_str().to_string()),
-        _ => return Err("Unexpected left expression in binary operation".to_string()),
-    };
-
-    let right = match right_expr.as_rule() {
-        Rule::identifier => Expression::Variable(right_expr.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_expr.as_str().to_string()),
-        _ => return Err("Unexpected right expression in binary operation".to_string()),
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-// ─── Asset Lookup Parsing ──────────────────────────────────────────────────────
-
-/// Parse asset_lookup_comparison: asset_lookup op (arith_expr | asset_lookup | identifier | literal)
-fn parse_asset_lookup_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left asset lookup")?;
-    let left = parse_asset_lookup_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::asset_lookup_arith_expr => parse_arith_expr_to_expression(right_pair)?,
-        Rule::asset_lookup => parse_asset_lookup_to_expression(right_pair)?,
-        Rule::bin2num_func => parse_bin2num(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in asset lookup comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse a standalone asset_lookup (not in a comparison context)
-fn parse_standalone_asset_lookup(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_asset_lookup_to_expression(pair)?;
-    // Wrap in a dummy comparison for standalone usage
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse an asset-id txid operand (a bytes32 identifier).
-fn parse_asset_id_txid(pair: Pair<Rule>) -> Expression {
-    Expression::Variable(pair.as_str().to_string())
-}
-
-/// Parse an asset-id gidx operand (an int identifier or a numeric literal).
-fn parse_asset_id_gidx(pair: Pair<Rule>) -> Expression {
-    match pair.as_rule() {
-        Rule::number_literal => Expression::Literal(pair.as_str().to_string()),
-        _ => Expression::Variable(pair.as_str().to_string()),
-    }
-}
-
-/// Parse the two Asset ID operands from an asset-group access pair.
-fn parse_asset_group_id_operands(pair: Pair<Rule>) -> Result<(Expression, Expression), String> {
-    let operand_parent = match pair.as_rule() {
-        Rule::tx_property_access => {
-            let mut inner = pair.into_inner();
-            let body = inner
-                .next()
-                .ok_or("asset id requires (txid, gidx) operands")?;
-            if body.as_rule() != Rule::tx_property_body || inner.next().is_some() {
-                return Err("asset id requires (txid, gidx) operands".to_string());
-            }
-            body
-        }
-        Rule::tx_property_body | Rule::asset_group_access => pair,
-        rule => return Err(format!("unexpected asset group operand parent: {rule:?}")),
-    };
-
-    let mut operands = operand_parent.into_inner();
-    let txid_pair = operands
-        .next()
-        .ok_or("asset id requires (txid, gidx) operands")?;
-    let gidx_pair = operands
-        .next()
-        .ok_or("asset id requires (txid, gidx) operands")?;
-
-    if txid_pair.as_rule() != Rule::identifier
-        || !matches!(gidx_pair.as_rule(), Rule::identifier | Rule::number_literal)
-        || operands.next().is_some()
-    {
-        return Err("asset id requires (txid, gidx) operands".to_string());
-    }
-
-    Ok((
-        parse_asset_id_txid(txid_pair),
-        parse_asset_id_gidx(gidx_pair),
-    ))
-}
-
-/// Shared parse for `tx.{inputs,outputs}[i].assets.{lookup,has}(txid, gidx)`:
-/// returns the source, the input/output index, and the two Asset ID operands.
-fn parse_asset_lookup_operands(
-    pair: Pair<Rule>,
-) -> Result<(AssetLookupSource, Expression, Expression, Expression), String> {
-    let mut inner = pair.into_inner();
-
-    // Parse source: "inputs" or "outputs"
-    let source_pair = inner.next().ok_or("Missing asset lookup source")?;
-    let source = match source_pair.as_str() {
-        "inputs" => AssetLookupSource::Input,
-        "outputs" => AssetLookupSource::Output,
-        _ => {
-            return Err(format!(
-                "Invalid asset lookup source: {}",
-                source_pair.as_str()
-            ))
-        }
-    };
-
-    // Parse array access index
-    let array_access = inner.next().ok_or("Missing array index")?;
-    let index_pair = array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing index value")?;
-    let index = match index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(index_pair.as_str().to_string()),
-        _ => Expression::Literal(index_pair.as_str().to_string()),
-    };
-
-    // Parse the canonical Asset ID operands: txid (bytes32) then gidx (int).
-    let asset_txid = parse_asset_id_txid(inner.next().ok_or("Missing asset txid")?);
-    let asset_gidx = parse_asset_id_gidx(inner.next().ok_or("Missing asset gidx")?);
-
-    Ok((source, index, asset_txid, asset_gidx))
-}
-
-/// Parse an asset_lookup pair into an Expression::AssetLookup
-fn parse_asset_lookup_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let (source, index, asset_txid, asset_gidx) = parse_asset_lookup_operands(pair)?;
-    Ok(Expression::AssetLookup {
-        source,
-        index: Box::new(index),
-        asset_txid: Box::new(asset_txid),
-        asset_gidx: Box::new(asset_gidx),
-    })
-}
-
-/// Parse an asset_has pair into an Expression::AssetHas
-fn parse_asset_has_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let (source, index, asset_txid, asset_gidx) = parse_asset_lookup_operands(pair)?;
-    Ok(Expression::AssetHas {
-        source,
-        index: Box::new(index),
-        asset_txid: Box::new(asset_txid),
-        asset_gidx: Box::new(asset_gidx),
-    })
-}
-
-/// Parse an asset_count pair into an Expression::AssetCount
-/// tx.inputs[i].assets.length or tx.outputs[o].assets.length
-fn parse_asset_count_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // Parse source: "inputs" or "outputs"
-    let source_pair = inner.next().ok_or("Missing asset count source")?;
-    let source = match source_pair.as_str() {
-        "inputs" => AssetLookupSource::Input,
-        "outputs" => AssetLookupSource::Output,
-        _ => {
-            return Err(format!(
-                "Invalid asset count source: {}",
-                source_pair.as_str()
-            ))
-        }
-    };
-
-    // Parse array access index
-    let array_access = inner.next().ok_or("Missing array index")?;
-    let index_pair = array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing index value")?;
-    let index = match index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(index_pair.as_str().to_string()),
-        _ => Expression::Literal(index_pair.as_str().to_string()),
-    };
-
-    Ok(Expression::AssetCount {
-        source,
-        index: Box::new(index),
-    })
-}
-
-/// Parse an asset_at pair into an Expression::AssetAt
-/// tx.inputs[i].assets[t].assetId or tx.outputs[o].assets[t].amount
-fn parse_asset_at_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // Parse source: "inputs" or "outputs"
-    let source_pair = inner.next().ok_or("Missing asset at source")?;
-    let source = match source_pair.as_str() {
-        "inputs" => AssetLookupSource::Input,
-        "outputs" => AssetLookupSource::Output,
-        _ => return Err(format!("Invalid asset at source: {}", source_pair.as_str())),
-    };
-
-    // Parse first array access (io_index)
-    let io_array_access = inner.next().ok_or("Missing io array index")?;
-    let io_index_pair = io_array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing io index value")?;
-    let io_index = match io_index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(io_index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(io_index_pair.as_str().to_string()),
-        _ => Expression::Literal(io_index_pair.as_str().to_string()),
-    };
-
-    // Parse second array access (asset_index)
-    let asset_array_access = inner.next().ok_or("Missing asset array index")?;
-    let asset_index_pair = asset_array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing asset index value")?;
-    let asset_index = match asset_index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(asset_index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(asset_index_pair.as_str().to_string()),
-        _ => Expression::Literal(asset_index_pair.as_str().to_string()),
-    };
-
-    // Parse property: "assetId" or "amount"
-    let property = inner
-        .next()
-        .ok_or("Missing asset property")?
-        .as_str()
-        .to_string();
-
-    Ok(Expression::AssetAt {
-        source,
-        io_index: Box::new(io_index),
-        asset_index: Box::new(asset_index),
-        property,
-    })
-}
-
-/// Parse a standalone asset_count (not in a comparison context)
-fn parse_standalone_asset_count(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_asset_count_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse a standalone asset_at (not in a comparison context)
-fn parse_standalone_asset_at(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_asset_at_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse asset_count_comparison: asset_count op expression
-fn parse_asset_count_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left asset count")?;
-    let left = parse_asset_count_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in asset count comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse a standalone asset_has (require-bare): leaves the presence flag.
-fn parse_standalone_asset_has(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_asset_has_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse asset_has_comparison: asset_has op (identifier | number_literal)
-fn parse_asset_has_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left asset has")?;
-    let left = parse_asset_has_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in asset has comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse a group_control_is pair: `group.controlIs(txid, gidx)` → GroupControlIs.
-fn parse_group_control_is_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let group = inner
-        .next()
-        .ok_or("Missing group in controlIs")?
-        .as_str()
-        .to_string();
-    let asset_txid = parse_asset_id_txid(inner.next().ok_or("Missing controlIs txid")?);
-    let asset_gidx = parse_asset_id_gidx(inner.next().ok_or("Missing controlIs gidx")?);
-    Ok(Expression::GroupControlIs {
-        group,
-        asset_txid: Box::new(asset_txid),
-        asset_gidx: Box::new(asset_gidx),
-    })
-}
-
-/// Parse a standalone group_control_is (require-bare): leaves the boolean.
-fn parse_standalone_group_control_is(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_group_control_is_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse group_control_is_comparison: group.controlIs(...) op (identifier | number_literal)
-fn parse_group_control_is_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left controlIs")?;
-    let left = parse_group_control_is_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in controlIs comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse asset_at_comparison: asset_at op expression
-fn parse_asset_at_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left asset at")?;
-    let left = parse_asset_at_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::asset_at => parse_asset_at_to_expression(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in asset at comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-// ─── Transaction Introspection Parsing ─────────────────────────────────────────
-
-/// Parse tx_introspection pair into an Expression::TxIntrospection
-fn parse_tx_introspection_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // Parse the property
-    let property = inner
-        .next()
-        .ok_or("Missing tx introspection property")?
-        .as_str()
-        .to_string();
-
-    Ok(Expression::TxIntrospection { property })
-}
-
-/// Parse a standalone tx_introspection (not in a comparison context)
-fn parse_standalone_tx_introspection(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_tx_introspection_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse tx_introspection_comparison: tx_introspection op expression
-fn parse_tx_introspection_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left tx introspection")?;
-    let left = parse_tx_introspection_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in tx introspection comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-// ─── Input/Output Introspection Parsing ─────────────────────────────────────────
-
-/// Parse input_introspection pair into an Expression::InputIntrospection
-/// tx.inputs[i].value, tx.inputs[i].scriptPubKey, etc.
-fn parse_input_introspection_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // Parse array access (the index)
-    let array_access = inner.next().ok_or("Missing input index")?;
-    let index_pair = array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing index value")?;
-    let index = match index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(index_pair.as_str().to_string()),
-        _ => Expression::Literal(index_pair.as_str().to_string()),
-    };
-
-    // Parse the property
-    let property = inner
-        .next()
-        .ok_or("Missing input introspection property")?
-        .as_str()
-        .to_string();
-
-    Ok(Expression::InputIntrospection {
-        index: Box::new(index),
-        property,
-    })
-}
-
-/// Parse output_introspection pair into an Expression::OutputIntrospection
-/// tx.outputs[o].value, tx.outputs[o].scriptPubKey, tx.outputs[o].nonce
-fn parse_output_introspection_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // Parse array access (the index)
-    let array_access = inner.next().ok_or("Missing output index")?;
-    let index_pair = array_access
-        .into_inner()
-        .next()
-        .ok_or("Missing index value")?;
-    let index = match index_pair.as_rule() {
-        Rule::number_literal => Expression::Literal(index_pair.as_str().to_string()),
-        Rule::identifier => Expression::Variable(index_pair.as_str().to_string()),
-        _ => Expression::Literal(index_pair.as_str().to_string()),
-    };
-
-    // Parse the property
-    let property = inner
-        .next()
-        .ok_or("Missing output introspection property")?
-        .as_str()
-        .to_string();
-
-    Ok(Expression::OutputIntrospection {
-        index: Box::new(index),
-        property,
-    })
-}
-
-/// Parse a standalone input_introspection (not in a comparison context)
-fn parse_standalone_input_introspection(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_input_introspection_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse a standalone output_introspection (not in a comparison context)
-fn parse_standalone_output_introspection(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_output_introspection_to_expression(pair)?;
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse input_introspection_comparison: input_introspection op expression
-fn parse_input_introspection_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left input introspection")?;
-    let left = parse_input_introspection_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::substr_func => parse_substr(right_pair)?,
-        Rule::input_introspection => parse_input_introspection_to_expression(right_pair)?,
-        Rule::output_introspection => parse_output_introspection_to_expression(right_pair)?,
-        Rule::tx_property_access | Rule::this_property_access => {
-            parse_tx_property_to_expression(right_pair)
-        }
-        Rule::constructor => parse_constructor_to_expression(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in input introspection comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse output_introspection_comparison: output_introspection op expression
-fn parse_output_introspection_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left output introspection")?;
-    let left = parse_output_introspection_to_expression(left_pair)?;
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::substr_func => parse_substr(right_pair)?,
-        Rule::input_introspection => parse_input_introspection_to_expression(right_pair)?,
-        Rule::output_introspection => parse_output_introspection_to_expression(right_pair)?,
-        Rule::tx_property_access | Rule::this_property_access => {
-            parse_tx_property_to_expression(right_pair)
-        }
-        Rule::constructor => parse_constructor_to_expression(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in output introspection comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-/// Parse an arithmetic expression in asset lookup context (e.g., lookup + amount)
-fn parse_arith_expr_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    let left_pair = inner.next().ok_or("Missing left operand")?;
-    let left = match left_pair.as_rule() {
-        Rule::asset_lookup => parse_asset_lookup_to_expression(left_pair)?,
-        Rule::identifier => Expression::Variable(left_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(left_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected left operand in arithmetic: {:?}",
-                left_pair.as_rule()
-            ))
-        }
-    };
-
-    let op = inner
-        .next()
-        .ok_or("Missing arithmetic operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right operand")?;
-    let right = match right_pair.as_rule() {
-        Rule::asset_lookup => parse_asset_lookup_to_expression(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right operand in arithmetic: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    Ok(Expression::BinaryOp {
-        left: Box::new(left),
-        op,
-        right: Box::new(right),
-    })
-}
-
-// ─── Asset Group Parsing ───────────────────────────────────────────────────────
-
-/// Parse asset_group_access: tx.assetGroups.find(id), tx.assetGroups.length,
-/// tx.assetGroups[k].property
-fn parse_asset_group_access(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let text = pair.as_str();
-
-    // Determine which variant of asset group access
-    if text.contains(".find(") {
-        // tx.assetGroups.find(txid, gidx)
-        let (asset_txid, asset_gidx) = parse_asset_group_id_operands(pair)?;
-        Ok(Requirement::Comparison {
-            left: Expression::GroupFind {
-                asset_txid: Box::new(asset_txid),
-                asset_gidx: Box::new(asset_gidx),
-            },
-            op: "==".to_string(),
-            right: Expression::Literal("true".to_string()),
-        })
-    } else if text.contains(".has(") {
-        // tx.assetGroups.has(txid, gidx)
-        let (asset_txid, asset_gidx) = parse_asset_group_id_operands(pair)?;
-        Ok(Requirement::Comparison {
-            left: Expression::GroupHas {
-                asset_txid: Box::new(asset_txid),
-                asset_gidx: Box::new(asset_gidx),
-            },
-            op: "==".to_string(),
-            right: Expression::Literal("true".to_string()),
-        })
-    } else if text.contains(".length") {
-        // tx.assetGroups.length
-        Ok(Requirement::Comparison {
-            left: Expression::AssetGroupsLength,
-            op: "==".to_string(),
-            right: Expression::Literal("true".to_string()),
-        })
-    } else {
-        // tx.assetGroups[k].property
-        let mut inner = pair.into_inner();
-        let array_access = inner.next().ok_or("Missing group index")?;
-        let index_pair = array_access
-            .into_inner()
-            .next()
-            .ok_or("Missing index value")?;
-        let index = match index_pair.as_rule() {
-            Rule::number_literal => Expression::Literal(index_pair.as_str().to_string()),
-            Rule::identifier => Expression::Variable(index_pair.as_str().to_string()),
-            _ => Expression::Literal(index_pair.as_str().to_string()),
-        };
-
-        let property = inner
-            .next()
-            .ok_or("Missing group property")?
-            .as_str()
-            .to_string();
-
-        let expr = match property.as_str() {
-            "sumInputs" => Expression::GroupSum {
-                index: Box::new(index),
-                source: GroupSumSource::Inputs,
-            },
-            "sumOutputs" => Expression::GroupSum {
-                index: Box::new(index),
-                source: GroupSumSource::Outputs,
-            },
-            "numInputs" => Expression::GroupNumIO {
-                index: Box::new(index),
-                source: GroupIOSource::Inputs,
-            },
-            "numOutputs" => Expression::GroupNumIO {
-                index: Box::new(index),
-                source: GroupIOSource::Outputs,
-            },
-            _ => Expression::GroupProperty {
-                group: format!("assetGroups[{}]", index_pair_to_string(&index)),
-                property,
-            },
-        };
-
-        Ok(Requirement::Comparison {
-            left: expr,
-            op: "==".to_string(),
-            right: Expression::Literal("true".to_string()),
-        })
-    }
-}
-
-/// Parse group_property_comparison: variable.property op expression
-fn parse_group_property_comparison(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-
-    let group_name = inner
-        .next()
-        .ok_or("Missing group variable name")?
-        .as_str()
-        .to_string();
-
-    let property = inner
-        .next()
-        .ok_or("Missing group property")?
-        .as_str()
-        .to_string();
-
-    let op = inner
-        .next()
-        .ok_or("Missing comparison operator")?
-        .as_str()
-        .to_string();
-
-    let right_pair = inner.next().ok_or("Missing right side expression")?;
-    let right = match right_pair.as_rule() {
-        Rule::group_property_arith_expr => {
-            // Parse group.property +/- value (e.g., tokenGroup.sumOutputs + amount)
-            let mut arith_inner = right_pair.into_inner();
-            let prop_access = arith_inner
-                .next()
-                .ok_or("Missing property access in arithmetic expression")?;
-            let mut prop_inner = prop_access.into_inner();
-            let var_name = prop_inner
-                .next()
-                .ok_or("Missing variable name in property access")?
-                .as_str()
-                .to_string();
-            let prop_name = prop_inner
-                .next()
-                .ok_or("Missing property name in property access")?
-                .as_str()
-                .to_string();
-            let left_expr = Expression::GroupProperty {
-                group: var_name,
-                property: prop_name,
-            };
-            let arith_op = arith_inner
-                .next()
-                .ok_or("Missing arithmetic operator")?
-                .as_str()
-                .to_string();
-            let right_operand = arith_inner
-                .next()
-                .ok_or("Missing right operand in arithmetic")?;
-            let right_expr = match right_operand.as_rule() {
-                Rule::bin2num_func => parse_bin2num(right_operand)?,
-                Rule::identifier => Expression::Variable(right_operand.as_str().to_string()),
-                Rule::number_literal => Expression::Literal(right_operand.as_str().to_string()),
-                _ => Expression::Property(right_operand.as_str().to_string()),
-            };
-            Expression::BinaryOp {
-                left: Box::new(left_expr),
-                op: arith_op,
-                right: Box::new(right_expr),
-            }
-        }
-        Rule::asset_lookup => parse_asset_lookup_to_expression(right_pair)?,
-        Rule::asset_group_access => {
-            // Parse the group access and extract the expression
-            let req = parse_asset_group_access(right_pair)?;
-            if let Requirement::Comparison { left, .. } = req {
-                left
-            } else {
-                return Err("Expected expression from asset group access".to_string());
-            }
-        }
-        Rule::identifier_property_access => {
-            // Parse variable.property (e.g., group.sumInputs)
-            let mut prop_inner = right_pair.into_inner();
-            let var_name = prop_inner
-                .next()
-                .ok_or("Missing variable name in property access")?
-                .as_str()
-                .to_string();
-            let prop_name = prop_inner
-                .next()
-                .ok_or("Missing property name in property access")?
-                .as_str()
-                .to_string();
-            Expression::GroupProperty {
-                group: var_name,
-                property: prop_name,
-            }
-        }
-        Rule::bin2num_func => parse_bin2num(right_pair)?,
-        Rule::identifier => Expression::Variable(right_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(right_pair.as_str().to_string()),
-        _ => {
-            return Err(format!(
-                "Unexpected right side in group property comparison: {:?}",
-                right_pair.as_rule()
-            ))
-        }
-    };
-
-    let left = Expression::GroupProperty {
-        group: group_name,
-        property,
-    };
-
-    Ok(Requirement::Comparison { left, op, right })
-}
-
-// ─── Streaming SHA256 Parsing ──────────────────────────────────────────
-
-/// Parse sha256Initialize(data) → Expression::Sha256Initialize
-fn parse_sha256_initialize(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let data_pair = inner.next().ok_or("Missing data in sha256Initialize")?;
-    let data = match data_pair.as_rule() {
-        Rule::identifier => Expression::Variable(data_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(data_pair.as_str().to_string()),
-        _ => Expression::Property(data_pair.as_str().to_string()),
-    };
-    Ok(Expression::Sha256Initialize {
-        data: Box::new(data),
-    })
-}
-
-/// Parse sha256Update(ctx, chunk) → Expression::Sha256Update
-fn parse_sha256_update(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let ctx_pair = inner.next().ok_or("Missing context in sha256Update")?;
-    let context = Expression::Variable(ctx_pair.as_str().to_string());
-
-    let chunk_pair = inner.next().ok_or("Missing chunk in sha256Update")?;
-    let chunk = match chunk_pair.as_rule() {
-        Rule::identifier => Expression::Variable(chunk_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(chunk_pair.as_str().to_string()),
-        _ => Expression::Property(chunk_pair.as_str().to_string()),
-    };
-    Ok(Expression::Sha256Update {
-        context: Box::new(context),
-        chunk: Box::new(chunk),
-    })
-}
-
-/// Parse sha256Finalize(ctx, lastChunk) → Expression::Sha256Finalize
-fn parse_sha256_finalize(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let ctx_pair = inner.next().ok_or("Missing context in sha256Finalize")?;
-    let context = Expression::Variable(ctx_pair.as_str().to_string());
-
-    let chunk_pair = inner.next().ok_or("Missing lastChunk in sha256Finalize")?;
-    let last_chunk = match chunk_pair.as_rule() {
-        Rule::identifier => Expression::Variable(chunk_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(chunk_pair.as_str().to_string()),
-        _ => Expression::Property(chunk_pair.as_str().to_string()),
-    };
-    Ok(Expression::Sha256Finalize {
-        context: Box::new(context),
-        last_chunk: Box::new(last_chunk),
-    })
-}
-
-// ─── Conversion & Arithmetic Parsing ───────────────────────────────────
-
-/// Parse neg64(value) → Expression::Neg64
-fn parse_neg64(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let value_pair = inner.next().ok_or("Missing value in neg64")?;
-    let value = match value_pair.as_rule() {
-        Rule::identifier => Expression::Variable(value_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(value_pair.as_str().to_string()),
-        _ => Expression::Property(value_pair.as_str().to_string()),
-    };
-    Ok(Expression::Neg64 {
-        value: Box::new(value),
-    })
-}
-
-/// Parse le64ToScriptNum(value) → Expression::Le64ToScriptNum
-fn parse_le64_to_script_num(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let value_pair = inner.next().ok_or("Missing value in le64ToScriptNum")?;
-    let value = match value_pair.as_rule() {
-        Rule::identifier => Expression::Variable(value_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(value_pair.as_str().to_string()),
-        _ => Expression::Property(value_pair.as_str().to_string()),
-    };
-    Ok(Expression::Le64ToScriptNum {
-        value: Box::new(value),
-    })
-}
-
-/// Parse le32ToLe64(value) → Expression::Le32ToLe64
-fn parse_le32_to_le64(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let value_pair = inner.next().ok_or("Missing value in le32ToLe64")?;
-    let value = match value_pair.as_rule() {
-        Rule::identifier => Expression::Variable(value_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(value_pair.as_str().to_string()),
-        _ => Expression::Property(value_pair.as_str().to_string()),
-    };
-    Ok(Expression::Le32ToLe64 {
-        value: Box::new(value),
-    })
-}
-
-// ─── Crypto Opcodes Parsing ────────────────────────────────────────────
-
-/// Parse ecMulScalarVerify(k, P, Q) → Expression::EcMulScalarVerify
-fn parse_ec_mul_scalar_verify(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    let scalar_pair = inner
-        .next()
-        .ok_or("Missing scalar k in ecMulScalarVerify")?;
-    let scalar = match scalar_pair.as_rule() {
-        Rule::identifier => Expression::Variable(scalar_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(scalar_pair.as_str().to_string()),
-        _ => Expression::Property(scalar_pair.as_str().to_string()),
-    };
-
-    let point_p_pair = inner.next().ok_or("Missing point P in ecMulScalarVerify")?;
-    let point_p = match point_p_pair.as_rule() {
-        Rule::identifier => Expression::Variable(point_p_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(point_p_pair.as_str().to_string()),
-        _ => Expression::Property(point_p_pair.as_str().to_string()),
-    };
-
-    let point_q_pair = inner.next().ok_or("Missing point Q in ecMulScalarVerify")?;
-    let point_q = match point_q_pair.as_rule() {
-        Rule::identifier => Expression::Variable(point_q_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(point_q_pair.as_str().to_string()),
-        _ => Expression::Property(point_q_pair.as_str().to_string()),
-    };
-
-    Ok(Expression::EcMulScalarVerify {
-        scalar: Box::new(scalar),
-        point_p: Box::new(point_p),
-        point_q: Box::new(point_q),
-    })
-}
-
-/// Parse tweakVerify(P, k, Q) → Expression::TweakVerify
-fn parse_tweak_verify(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    let point_p_pair = inner.next().ok_or("Missing point P in tweakVerify")?;
-    let point_p = match point_p_pair.as_rule() {
-        Rule::identifier => Expression::Variable(point_p_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(point_p_pair.as_str().to_string()),
-        _ => Expression::Property(point_p_pair.as_str().to_string()),
-    };
-
-    let tweak_pair = inner.next().ok_or("Missing tweak k in tweakVerify")?;
-    let tweak = match tweak_pair.as_rule() {
-        Rule::identifier => Expression::Variable(tweak_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(tweak_pair.as_str().to_string()),
-        _ => Expression::Property(tweak_pair.as_str().to_string()),
-    };
-
-    let point_q_pair = inner.next().ok_or("Missing point Q in tweakVerify")?;
-    let point_q = match point_q_pair.as_rule() {
-        Rule::identifier => Expression::Variable(point_q_pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(point_q_pair.as_str().to_string()),
-        _ => Expression::Property(point_q_pair.as_str().to_string()),
-    };
-
-    Ok(Expression::TweakVerify {
-        point_p: Box::new(point_p),
-        tweak: Box::new(tweak),
-        point_q: Box::new(point_q),
-    })
-}
-
-/// Parse checkSigFromStackVerify(sig, pubkey, msg) → Requirement::CheckSig (verify variant)
-fn parse_check_sig_from_stack_verify(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let mut inner = pair.into_inner();
-    let signature = inner
-        .next()
-        .ok_or("Missing signature in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-    let pubkey = inner
-        .next()
-        .ok_or("Missing pubkey in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-    let message = inner
-        .next()
-        .ok_or("Missing message in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-
-    Ok(Requirement::Comparison {
-        left: Expression::CheckSigFromStackVerify {
-            signature,
-            pubkey,
-            message,
-        },
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Parse checkSigFromStackVerify for primary expression context
-fn parse_check_sig_from_stack_verify_expr(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let signature = inner
-        .next()
-        .ok_or("Missing signature in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-    let pubkey = inner
-        .next()
-        .ok_or("Missing pubkey in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-    let message = inner
-        .next()
-        .ok_or("Missing message in checkSigFromStackVerify")?
-        .as_str()
-        .to_string();
-
-    Ok(Expression::CheckSigFromStackVerify {
-        signature,
-        pubkey,
-        message,
-    })
-}
-
-// ─── Byte-string Manipulation Parsing ──────────────────────────────────
-
-/// Helper: convert an inner pair to an Expression for the byte-string
-/// and packet primitives. Identifiers → Variable, numbers → Literal,
-/// everything else → Property.
-fn parse_atom_pair(pair: Pair<Rule>) -> Expression {
-    match pair.as_rule() {
-        Rule::identifier => Expression::Variable(pair.as_str().to_string()),
-        Rule::number_literal => Expression::Literal(pair.as_str().to_string()),
-        _ => Expression::Property(pair.as_str().to_string()),
-    }
-}
-
-/// Parse a `byte_value` rule into an Expression. Used wherever the grammar
-/// accepts an arbitrary byte-producing operand (substr/cat/bin2num/size args).
-fn parse_byte_value(pair: Pair<Rule>) -> Result<Expression, String> {
-    // byte_value wraps exactly one inner rule.
-    let inner = pair.into_inner().next().ok_or("Empty byte_value")?;
-    match inner.as_rule() {
-        Rule::substr_func => parse_substr(inner),
-        Rule::cat_func => parse_cat(inner),
-        Rule::num2bin_func => parse_num2bin(inner),
-        Rule::packet_inspect => parse_packet_inspect(inner),
-        Rule::input_packet_inspect => parse_input_packet_inspect(inner),
-        Rule::input_introspection => parse_input_introspection_to_expression(inner),
-        Rule::output_introspection => parse_output_introspection_to_expression(inner),
-        Rule::asset_at => parse_asset_at_to_expression(inner),
-        Rule::identifier => Ok(Expression::Variable(inner.as_str().to_string())),
-        r => Err(format!("Unsupported byte_value rule: {:?}", r)),
-    }
-}
-
-/// Parse substr(data, offset, size) → Expression::Substr
-fn parse_substr(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let data = parse_byte_value(inner.next().ok_or("Missing data in substr")?)?;
-    let offset = parse_atom_pair(inner.next().ok_or("Missing offset in substr")?);
-    let size = parse_atom_pair(inner.next().ok_or("Missing size in substr")?);
-    Ok(Expression::Substr {
-        data: Box::new(data),
-        offset: Box::new(offset),
-        size: Box::new(size),
-    })
-}
-
-/// Parse cat(a, b) → Expression::Cat
-fn parse_cat(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let left = parse_byte_value(inner.next().ok_or("Missing first argument in cat")?)?;
-    let right = parse_byte_value(inner.next().ok_or("Missing second argument in cat")?)?;
-    Ok(Expression::Cat {
-        left: Box::new(left),
-        right: Box::new(right),
-    })
-}
-
-/// Parse bin2num(data) → Expression::Bin2Num
-fn parse_bin2num(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let data = parse_byte_value(inner.next().ok_or("Missing data in bin2num")?)?;
-    Ok(Expression::Bin2Num {
-        data: Box::new(data),
-    })
-}
-
-/// Parse num2bin(value, size) → Expression::Num2Bin
-fn parse_num2bin(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let value = parse_atom_pair(inner.next().ok_or("Missing value in num2bin")?);
-    let size = parse_atom_pair(inner.next().ok_or("Missing size in num2bin")?);
-    Ok(Expression::Num2Bin {
-        value: Box::new(value),
-        size: Box::new(size),
-    })
-}
-
-/// Parse size(data) → Expression::SizeOf
-fn parse_size(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let data = parse_byte_value(inner.next().ok_or("Missing data in size")?)?;
-    Ok(Expression::SizeOf {
-        data: Box::new(data),
-    })
-}
-
-/// Parse tx.packet(packetType) → Expression::PacketInspect
-fn parse_packet_inspect(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-    let packet_type = parse_atom_pair(inner.next().ok_or("Missing packet type in tx.packet()")?);
-    Ok(Expression::PacketInspect {
-        packet_type: Box::new(packet_type),
-    })
-}
-
-/// Parse tx.inputs[i].packet(packetType) → Expression::InputPacketInspect
-fn parse_input_packet_inspect(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // First child: array_access — extract the index expression
-    let array_access = inner
-        .next()
-        .ok_or("Missing input index in tx.inputs[i].packet()")?;
-    let index_pair = array_access
-        .into_inner()
-        .next()
-        .ok_or("Empty array access in tx.inputs[i].packet()")?;
-    let index = parse_atom_pair(index_pair);
-
-    let packet_type = parse_atom_pair(
-        inner
-            .next()
-            .ok_or("Missing packet type in tx.inputs[i].packet()")?,
-    );
-
-    Ok(Expression::InputPacketInspect {
-        index: Box::new(index),
-        packet_type: Box::new(packet_type),
-    })
-}
-
-// ─── Constructor Parsing ───────────────────────────────────────────────────────
-
-/// Parse a `constructor` rule pair into an `Expression::ContractInstance`.
-///
-/// Handles `new ContractName(arg1, arg2, ...)` and produces
-/// `ContractInstance { contract_name, args }` which the compiler lowers to
-/// a `<VTXO:ContractName(...)>` scriptPubKey placeholder.
-fn parse_constructor_to_expression(pair: Pair<Rule>) -> Result<Expression, String> {
-    let mut inner = pair.into_inner();
-
-    // First child: contract name identifier
-    let contract_name = inner
-        .next()
-        .ok_or("Parse error: Missing contract name in constructor")?
-        .as_str()
-        .to_string();
-
-    // Second child (optional): constructor_args rule
-    let args = if let Some(args_pair) = inner.next() {
-        parse_constructor_args(args_pair)?
-    } else {
-        Vec::new()
-    };
-
-    Ok(Expression::ContractInstance {
-        contract_name,
-        args,
-    })
-}
-
-/// Parse constructor arguments into a Vec<Expression>.
-///
-/// `constructor_args` is a named (non-silent) rule whose children are the
-/// alternatives matched by the silent `complex_expression` rule — so we
-/// see the raw inner rules (identifier, number_literal, etc.) directly.
-fn parse_constructor_args(pair: Pair<Rule>) -> Result<Vec<Expression>, String> {
-    let mut args = Vec::new();
-
-    for inner in pair.into_inner() {
-        let expr = match inner.as_rule() {
-            Rule::identifier => Expression::Variable(inner.as_str().to_string()),
-            Rule::number_literal => Expression::Literal(inner.as_str().to_string()),
-            Rule::constructor => parse_constructor_to_expression(inner)?,
-            Rule::input_introspection => parse_input_introspection_to_expression(inner)?,
-            Rule::output_introspection => parse_output_introspection_to_expression(inner)?,
-            Rule::tx_introspection => parse_tx_introspection_to_expression(inner)?,
-            _ => {
-                // Fall back to treating as a variable/property reference
-                Expression::Variable(inner.as_str().to_string())
-            }
-        };
-        args.push(expr);
-    }
-
-    Ok(args)
-}
-
-// ─── Helper Functions ──────────────────────────────────────────────────────────
-
-/// Parse tx_property_access into the appropriate Expression type
-/// Handles special patterns like tx.assetGroups[idx].sumInputs/sumOutputs
-/// Reject malformed asset-API calls that fell through to the generic property
-/// path. A well-formed `.assets.lookup`/`.assets.has` matches the dedicated
-/// `asset_lookup`/`asset_has` rules (which require exactly two operands) before
-/// any property fallback, so seeing one of these method names in a property
-/// string means a legacy single-argument or otherwise malformed call.
-fn reject_malformed_asset_call(text: &str) -> Result<(), String> {
-    if text.contains(".assets.lookup(") {
-        return Err(format!(
-            "asset lookup requires two operands `lookup(txid, gidx)`: {text}"
-        ));
-    }
-    if text.contains(".assets.has(") {
-        return Err(format!(
-            "asset presence check requires two operands `has(txid, gidx)`: {text}"
-        ));
-    }
-    Ok(())
-}
-
-fn parse_tx_property_to_expr(pair: Pair<Rule>) -> Result<Expression, String> {
-    let text = pair.as_str();
-
-    reject_malformed_asset_call(text)?;
-
-    // Handle tx.assetGroups.find(txid, gidx)
-    if text.starts_with("tx.assetGroups.find(") && text.ends_with(')') {
-        let (asset_txid, asset_gidx) = parse_asset_group_id_operands(pair)?;
-        return Ok(Expression::GroupFind {
-            asset_txid: Box::new(asset_txid),
-            asset_gidx: Box::new(asset_gidx),
-        });
-    }
-
-    // Handle tx.assetGroups.has(txid, gidx)
-    if text.starts_with("tx.assetGroups.has(") && text.ends_with(')') {
-        let (asset_txid, asset_gidx) = parse_asset_group_id_operands(pair)?;
-        return Ok(Expression::GroupHas {
-            asset_txid: Box::new(asset_txid),
-            asset_gidx: Box::new(asset_gidx),
-        });
-    }
-
-    // Handle tx.assetGroups.length
-    if text == "tx.assetGroups.length" {
-        return Ok(Expression::AssetGroupsLength);
-    }
-
-    // Handle tx.assetGroups[idx].sumInputs or tx.assetGroups[idx].sumOutputs
-    if text.starts_with("tx.assetGroups[") {
-        if let Some(bracket_start) = text.find('[') {
-            if let Some(bracket_end) = text.find(']') {
-                let idx_str = &text[bracket_start + 1..bracket_end];
-                let index = if idx_str.chars().all(|c| c.is_ascii_digit()) {
-                    Expression::Literal(idx_str.to_string())
-                } else {
-                    Expression::Variable(idx_str.to_string())
-                };
-
-                if text.ends_with(".sumInputs") {
-                    return Ok(Expression::GroupSum {
-                        index: Box::new(index),
-                        source: GroupSumSource::Inputs,
-                    });
-                } else if text.ends_with(".sumOutputs") {
-                    return Ok(Expression::GroupSum {
-                        index: Box::new(index),
-                        source: GroupSumSource::Outputs,
-                    });
-                } else if text.ends_with(".numInputs") {
-                    return Ok(Expression::GroupNumIO {
-                        index: Box::new(index),
-                        source: GroupIOSource::Inputs,
-                    });
-                } else if text.ends_with(".numOutputs") {
-                    return Ok(Expression::GroupNumIO {
-                        index: Box::new(index),
-                        source: GroupIOSource::Outputs,
-                    });
-                }
-            }
-        }
-    }
-
-    // Handle tx.input.current
-    if text.starts_with("tx.input.current") {
-        let property = if text == "tx.input.current" {
-            None
-        } else {
-            text.strip_prefix("tx.input.current.")
-                .map(|rest| rest.to_string())
-        };
-        return Ok(Expression::CurrentInput(property));
-    }
-
-    // Default: treat as a property string
-    Ok(Expression::Property(text.to_string()))
-}
-
-/// Convert a tx_property_access or this_property_access pair into an Expression
-fn parse_tx_property_to_expression(pair: Pair<Rule>) -> Expression {
-    let property_access = pair.as_str().to_string();
-
-    // Special handling for tx.input.current
-    if property_access.starts_with("tx.input.current") {
-        let property = if property_access == "tx.input.current" {
-            None
-        } else {
-            let parts: Vec<&str> = property_access.split('.').collect();
-            if parts.len() >= 4 {
-                Some(parts[3].to_string())
-            } else {
-                None
-            }
-        };
-        Expression::CurrentInput(property)
-    } else {
-        Expression::Property(property_access)
-    }
-}
-
-/// Parse a tx_property_access or this_property_access as a standalone Requirement
-fn parse_property_access_as_requirement(pair: Pair<Rule>) -> Result<Requirement, String> {
-    let expr = parse_tx_property_to_expression(pair);
-    Ok(Requirement::Comparison {
-        left: expr,
-        op: "==".to_string(),
-        right: Expression::Literal("true".to_string()),
-    })
-}
-
-/// Helper to convert an Expression index to a string representation
-fn index_pair_to_string(expr: &Expression) -> String {
-    match expr {
-        Expression::Literal(s) => s.clone(),
-        Expression::Variable(s) => s.clone(),
-        _ => "?".to_string(),
-    }
-}
-
 /// Parse parameter list from contracts or functions
-fn parse_parameters(params: Pair<Rule>) -> Result<Vec<Parameter>, String> {
+pub(crate) fn parse_parameters(params: Pair<Rule>) -> Result<Vec<Parameter>, String> {
     let mut parameters = Vec::new();
     for param_pair in params.into_inner() {
         if param_pair.as_rule() == Rule::parameter {
